requirements.txt
tensorflow
numpy
scipy
scikit-learn
pandas
matplotlib
pillow